Converted logging middleware into intercept middleware
