+++ examples/app1/api/api.go
@@ -1,13 +1,12 @@
 package app1
 
 import (
-	"net/http"
-	"time"
-
 	"github.com/BSick7/go-api"
 	"github.com/BSick7/go-api/cors"
 	"github.com/BSick7/go-api/gzip"
+	"github.com/BSick7/go-api/intercept"
 	"github.com/BSick7/go-api/json"
+	"github.com/BSick7/go-api/jwt"
 	"github.com/BSick7/go-api/logging"
 	"github.com/BSick7/go-api/recovery"
 	"github.com/BSick7/go-api/standard"
@@ -15,18 +14,6 @@ import (
 )
 
 func Server() *api.Server {
-	loggingCfg := logging.Config{
-		Prefix:  "[app1]",
-		Log100s: false,
-		Log200s: false,
-		Log300s: false,
-		Log400s: false,
-		Log500s: false,
-		OnRequest: func(r *http.Request, data logging.ResponseData, duration time.Duration) {
-
-		},
-	}
-
 	apiServer := &api.Server{
 		Router: mux.NewRouter().
 			StrictSlash(false).
@@ -34,10 +21,11 @@ func Server() *api.Server {
 			UseEncodedPath(),
 	}
 	api.DefaultFallbackBehavior(apiServer)
-	apiServer.Use(logging.EndpointLoggerMiddleware(loggingCfg))
 	apiServer.Use(gzip.Middleware())
 	apiServer.Use(recovery.PanicMiddleware())
 	apiServer.Use(cors.Middleware(cors.DefaultSettings))
+	apiServer.Use(jwt.Middleware())
+	apiServer.Use(intercept.Middleware(logging.LogAllRequests("[app1] ")))
 	apiServer.Register(endpoints...)
 	apiServer.Register(cors.Preflight())
 	return apiServer
